refactor: 폴더와 중복된 파일 이름 변경
