+++ .eslintrc.json
@@ -1,5 +1,7 @@
 {
-  "rules": {},
+  "rules": {
+    "max-lines-per-function": ["error", { "max": 10 }]
+  },
   "env": {
     "es6": true,
     "node": true
@@ -8,8 +10,5 @@
     "ecmaVersion": "latest",
     "sourceType": "module"
   },
-  "extends": [
-    "eslint:recommended",
-    "plugin:prettier/recommended"
-  ]
+  "extends": ["eslint:recommended", "plugin:prettier/recommended"]
 }
